requirements.txt
torch
torchvision
pillow
tqdm
numpy
pytorch-fid
lpips
